Modularized Code to use friends.js as a controller
